Layout edited + Bugfixing.

+++ gui/gui.py
@@ -4,7 +4,7 @@ from nicegui import ui, app
 
 from gui.utils import validate_positive_integer
 from service import make_line_plot_service, get_all_heart_service, save_heart_service, save_plot_to_document, \
-    all_heart_values_as_json_service
+    all_values_as_json_service
 
 table = None
 plot = None
@@ -25,27 +25,22 @@ def save_values(diastolic_input, systolic_input, pulse_input, date=None, time=No
             full_datetime = f"{date.value} {time.value}"
             date = datetime.strptime(f"{full_datetime}",
                                      "%Y-%m-%d %H:%M")
-            # todo Plot aktualisieren
+
         if (validate_positive_integer(diastolic) or validate_positive_integer(systolic)
                 or validate_positive_integer(pulse)):
             ui.notify('Bitte nur positive Werte eingeben!', color='red')
             return
 
         save_heart_service(systolic, diastolic, pulse, date)
-        update_view(systolic, diastolic, pulse, date)
+        update_view()
         ui.notify(f'Diastolisch: {diastolic}, Systolisch: {systolic}, Puls: {pulse}', color='green')
     except Exception as e:
         raise e
 
 
-def update_view(systolic, diastolic, pulse, date):
+def update_view():
     global table, plot, raw_plot, no_data_label, no_data_icon
-    # todo
-    if table:
-        table.rows = all_heart_values_as_json_service()
-        ui.update(table)
-
-    save_heart_service(systolic, diastolic, pulse)
+    all_heart_values = get_all_heart_service()
 
     new_fig = make_line_plot_service(get_all_heart_service())
 
@@ -53,11 +48,18 @@ def update_view(systolic, diastolic, pulse, date):
         plot.figure = new_fig
         plot.figure.layout = new_fig.layout
         ui.update(plot)
-        no_data_label.delete()
-        no_data_icon.delete()
-        no_data_label = None
-        no_data_icon = None
 
+        if no_data_label and no_data_icon:
+            no_data_label.delete()
+            no_data_icon.delete()
+            no_data_label = None
+            no_data_icon = None
+
+        # todo
+    if table:
+        table.rows = all_values_as_json_service(all_heart_values)
+        table.update()
+        ui.update(table)
 
 
 def build_gui():
@@ -66,6 +68,7 @@ def build_gui():
     global raw_plot
     global no_data_label
     global no_data_icon
+    current_date = datetime
 
     with (ui.grid(columns=1).classes('justify-center items-center w-full')):
         with ui.tabs().classes('w-full') as tabs:
@@ -108,10 +111,11 @@ def build_gui():
                                                validation=validate_positive_integer).classes('w-full')
 
                         ui.label("Gib Datum und Uhrzeit ein:")
+                        date_input = ui.input("Datum", placeholder="Tag.Monat.Jahr", value=current_date.today().date()
+                                              .isoformat()).props('type=date').classes('w-full')
 
-                        date_input = ui.input("Datum", placeholder="Tag.Monat.Jahr").props('type=date').classes(
-                            'w-full')
-                        time_input = ui.input("Uhrzeit", placeholder="hh:mm").props('type=time').classes('w-full')
+                        time_input = ui.input("Uhrzeit", placeholder="hh:mm", value=current_date.now().
+                                              strftime('%H:%M')).props('type=time').classes('w-full')
 
                         ui.button('Werte speichern',
                                   on_click=lambda: save_values(diastolic_input, systolic_input, pulse_input,
@@ -124,7 +128,6 @@ def build_gui():
                           ).classes('px-6 py-2')
 
             with ui.tab_panel(three):
-                table = ui.table(rows=all_heart_values_as_json_service(), pagination=10,
+                table = ui.table(rows=all_values_as_json_service(all_heart_values), pagination=10,
                                  on_pagination_change=lambda e: ui.notify(e.value))
-
         ui.button('App schließen', on_click=lambda: app.shutdown()).classes('bg-red-500 text-white px-6 py-2')
